# app/src/main/java/com/demo/app/model/AccountAlias.java
// Generated by AI (GitHub Copilot)
// Author: Nisarg Raval
package com.demo.app.model;

import jakarta.persistence.*;

@Entity
@Table(name = "account_alias", uniqueConstraints = @UniqueConstraint(columnNames = {"alias"}))
public class AccountAlias {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false, unique = true)
    private String alias;

    @Column(nullable = false)
    private String status;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "user_id")
    private User user;

    // Getters and setters
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public String getAlias() { return alias; }
    public void setAlias(String alias) { this.alias = alias; }
    public String getStatus() { return status; }
    public void setStatus(String status) { this.status = status; }
    public User getUser() { return user; }
    public void setUser(User user) { this.user = user; }
}