# app/src/main/java/com/demo/app/model/User.java
// Generated by AI (GitHub Copilot)
// Author: Nisarg Raval
package com.demo.app.model;

import jakarta.persistence.*;
import java.util.Set;

@Entity
@Table(name = "users")
public class User {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(unique = true, nullable = false)
    private String email;

    @Column(nullable = false)
    private String password;

    private String name;

    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
    private Set<AccountAlias> accountAliases;

    // Getters and setters
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public Set<AccountAlias> getAccountAliases() { return accountAliases; }
    public void setAccountAliases(Set<AccountAlias> accountAliases) { this.accountAliases = accountAliases; }
}