# app/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: User Registration and Account Alias API
  version: 1.0.0
servers:
  - url: /api
paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User registered
        '400':
          description: User already registered
  /users/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
        '400':
          description: Invalid email or password
  /users/{id}:
    get:
      summary: Get user details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
        '404':
          description: Not found
    put:
      summary: Update user details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
  /aliases/create/{userId}:
    post:
      summary: Create account alias
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountAlias'
      responses:
        '200':
          description: Alias created
        '400':
          description: Alias already exists or user not found
  /aliases/{id}:
    get:
      summary: Get alias details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alias details
        '404':
          description: Not found
    put:
      summary: Update alias status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountAlias'
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
    delete:
      summary: Delete alias
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    UserLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    AccountAlias:
      type: object
      properties:
        alias:
          type: string
        status:
          type: string